# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GOEN API",
        "contact": {
            "name": "Jorge Soto",
            "email": "jorgemsm@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/grade": {
            "get": {
                "tags": [
                    "Model: Grade",
                    "Role: Instructor"
                ],
                "summary": "Display a listing of Grades",
                "operationId": "getAllGrades",
                "responses": {
                    "200": {
                        "description": "Grades found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GradeCollection"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Model: Grade",
                    "Role: Instructor"
                ],
                "summary": "Store grade",
                "operationId": "storeGrade",
                "parameters": [
                    {
                        "name": "score",
                        "in": "query",
                        "description": "Grade's score",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "module_id",
                        "in": "query",
                        "description": "Grade's module id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Grade's user id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns created grade",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleGrade"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/api/module": {
            "get": {
                "tags": [
                    "Model: Module"
                ],
                "summary": "Display a listing of Grades",
                "operationId": "getAllGrades",
                "responses": {
                    "200": {
                        "description": "Modules found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ModuleCollection"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Model: Module"
                ],
                "summary": "Store module",
                "operationId": "storeModule",
                "parameters": [
                    {
                        "name": "period_id",
                        "in": "query",
                        "description": "Module's period",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Module's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "section",
                        "in": "query",
                        "description": "Module's section",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "Module's price",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "schedule_id",
                        "in": "query",
                        "description": "Module's schedule",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns created module",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleModule"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Module already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/api/schedule": {
            "get": {
                "tags": [
                    "Model: Schedule"
                ],
                "summary": "Display a listing of Grades",
                "operationId": "getAllGrades",
                "responses": {
                    "200": {
                        "description": "Schedules found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ScheduleCollection"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Model: Schedule"
                ],
                "summary": "Store schedule",
                "operationId": "storeSchedule",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Class date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Class start time",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "until",
                        "in": "query",
                        "description": "Class end time",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns created schedule",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleSchedule"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Schedule already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/schedule/{schedule}": {
            "get": {
                "tags": [
                    "Model: Schedule"
                ],
                "summary": "Get a Schedule by Id",
                "operationId": "getScheduleById",
                "parameters": [
                    {
                        "name": "schedule",
                        "in": "path",
                        "description": "Id of schedule",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Schedule found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleSchedule"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Model: Schedule"
                ],
                "summary": "Update schedule data",
                "operationId": "updateSchedule",
                "parameters": [
                    {
                        "name": "schedule",
                        "in": "path",
                        "description": "Id of schedule",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Class start time",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "until",
                        "in": "query",
                        "description": "Class end time",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns updated schedule",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleSchedule"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Model: Schedule"
                ],
                "summary": "Delete schedule",
                "operationId": "deleteSchedule",
                "parameters": [
                    {
                        "name": "schedule",
                        "in": "path",
                        "description": "Id of schedule",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Schedule deleted"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Grade": {
                "title": "Grade",
                "description": "Grade model",
                "required": [
                    "id",
                    "score",
                    "module_id",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Unique identificator",
                        "type": "integer",
                        "format": "int64"
                    },
                    "score": {
                        "title": "score",
                        "description": "User's score in module",
                        "type": "number",
                        "format": "double"
                    },
                    "module_id": {
                        "title": "module_id",
                        "description": "Module in which the user got the score.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "User to whom the score belongs.",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
            "Module": {
                "title": "Module",
                "description": "Module model",
                "required": [
                    "id",
                    "name",
                    "section",
                    "period_id",
                    "schedule_id",
                    "price_id",
                    "instructor_id"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Unique identificator",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "title": "name",
                        "description": "Module name",
                        "type": "string"
                    },
                    "section": {
                        "title": "section",
                        "description": "Module section",
                        "type": "string"
                    },
                    "period_id": {
                        "title": "period_id",
                        "description": "Period to whom the module belongs",
                        "type": "integer",
                        "format": "int64"
                    },
                    "schedule_id": {
                        "title": "schedule_id",
                        "description": "Schedule where the module takes places.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "price_id": {
                        "title": "price_id",
                        "description": "Price to register in module.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "clan_id": {
                        "title": "clan_id",
                        "description": "If M0, the module needs a clan to give new students.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "instructor_id": {
                        "title": "instructor_id",
                        "description": "Instructor in charge of the module.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "assistant_id": {
                        "title": "assistant_id",
                        "description": "Instructor's assistant in the module.",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
...
            "Schedule": {
                "title": "Schedule",
                "description": "Modules possible schedule",
                "required": [
                    "id",
                    "start_date",
                    "from",
                    "until"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Unique identificator",
                        "type": "integer",
                        "format": "int64"
                    },
                    "start_date": {
                        "title": "start_date",
                        "description": "Day in which the class takes places. YYYY-MM-DD",
                        "type": "string",
                        "format": "date"
                    },
                    "from": {
                        "title": "from",
                        "description": "Start time of the class. HH:MM:SS",
                        "type": "string"
                    },
                    "until": {
                        "title": "until",
                        "description": "Time in which the class ends. HH:MM:SS",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "title": "error",
                        "description": "Error message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "SingleGrade": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Grade"
                    }
                },
                "type": "object"
            },
            "GradeCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Grade"
                        }
                    }
                },
                "type": "object"
            },
            "SingleModule": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Module"
                    }
                },
                "type": "object"
            },
            "ModuleCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Module"
                        }
                    }
                },
                "type": "object"
            },
...
            "SingleSchedule": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Schedule"
                    }
                },
                "type": "object"
            },
            "ScheduleCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Schedule"
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}